+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Audio Player project will be documented in this file.
 ### Changed
 - Event dispatcher for Add-ons
 - getID3 1.9.20-202109010614
+- remove non permitted characters #532
 
 ## 3.2.1 - 2021-08-20
 ### Fixed