+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
 All notable changes to the Audio Player project will be documented in this file.
 
-
+## 3.3.0 - 2021-09-xx
 ### Changed
 - Event dispatcher for Add-ons
 - getID3 1.9.20-202109010614
 - remove non permitted characters #532
+
+### Fixed
+- Type issue in scanner #528
 
 ## 3.2.1 - 2021-08-20
 ### Fixed